[Refact] 모임 관리 TS 전환 및 TanstackQuery 적용
